src/file/pyproject.toml
[project]
name = "file"
version = "2.0.1"
requires-python = ">=3.12"
dependencies = [
    "aiobotocore==2.20.0",
    "alembic==1.14.0",
    "asyncpg==0.30.0",
    "cashews==7.4.0",
    "greenlet>=3.2.3",
    "grpc-accesslog==0.3.0",
    "grpcio==1.69.0",
    "grpcio-tools==1.69.0",
    "inject>=5.3.0",
    "picologging==0.9.3",
    "py-async-grpc-prometheus==0.2.0",
    "redis==5.2.1",
    "sqlalchemy==2.0.37",
    "types-aiobotocore[s3]==2.20.0",
    "uvicorn[standard]==0.34.0",
]

[dependency-groups]
test = [
    "coverage==7.9.2",
    "pytest==8.3.4",
    "pytest-asyncio==0.25.3",
    "pytest-cov==6.0.0",
]
